+++ lib/punchblock/event/stopped_speaking.rb
@@ -1,0 +1,13 @@
+# encoding: utf-8
+
+require 'punchblock/event/active_speaker'
+
+module Punchblock
+  class Event
+    class StoppedSpeaking < Event
+      register :'stopped-speaking', :core
+
+      include ActiveSpeaker
+    end
+  end
+end