+++ lib/punchblock/translator/asterisk/component/output.rb
@@ -7,6 +7,7 @@ module Punchblock
     class Asterisk
       module Component
         class Output < Component
+          include StopByRedirect
 
           def setup
             @media_engine = @call.translator.media_engine
@@ -31,8 +32,11 @@ module Punchblock
                 case node
                 when RubySpeech::SSML::Audio
                   lambda { current_actor.play_audio! node.src }
+                when String
+                  return unrenderable_doc_error if node.include?(' ')
+                  lambda { current_actor.play_audio! node }
                 else
-                  return with_error 'unrenderable document error', 'The provided document could not be rendered.'
+                  return unrenderable_doc_error
                 end
               end.compact
 
@@ -53,18 +57,20 @@ module Punchblock
               end
             when :unimrcp
               send_ref
+              output_component = current_actor
               @call.send_agi_action! 'EXEC MRCPSynth', escaped_doc, mrcpsynth_options do |complete_event|
                 pb_logger.debug "MRCPSynth completed with #{complete_event}."
-                send_complete_event success_reason
+                output_component.send_complete_event! success_reason
               end
             when :swift
               doc = escaped_doc
               doc << "|1|1" if [:any, :dtmf].include? @component_node.interrupt_on
               doc.insert 0, "#{@component_node.voice}^" if @component_node.voice
               send_ref
+              output_component = current_actor
               @call.send_agi_action! 'EXEC Swift', doc do |complete_event|
                 pb_logger.debug "Swift completed with #{complete_event}."
-                send_complete_event success_reason
+                output_component.send_complete_event! success_reason
               end
             end
           end
@@ -92,6 +98,10 @@ module Punchblock
           end
 
           private
+
+          def unrenderable_doc_error
+            with_error 'unrenderable document error', 'The provided document could not be rendered.'
+          end
 
           def escaped_doc
             @component_node.ssml.to_s.squish.gsub(/["\\]/) { |m| "\\#{m}" }