+++ spec/punchblock/translator/asterisk_spec.rb
@@ -132,7 +132,7 @@ module Punchblock
         context "with an unknown call ID" do
           it 'sends an error in response to the command' do
             subject.execute_call_command command
-            command.response.should be == ProtocolError.new('call-not-found', "Could not find a call with ID #{call_id}", call_id, nil)
+            command.response.should be == ProtocolError.new.setup('call-not-found', "Could not find a call with ID #{call_id}", call_id, nil)
           end
         end
       end
@@ -161,7 +161,7 @@ module Punchblock
         context "with an unknown component ID" do
           it 'sends an error in response to the command' do
             subject.execute_component_command command
-            command.response.should be == ProtocolError.new('component-not-found', "Could not find a component with ID #{component_id}", nil, component_id)
+            command.response.should be == ProtocolError.new.setup('component-not-found', "Could not find a component with ID #{component_id}", nil, component_id)
           end
         end
       end
@@ -218,7 +218,7 @@ module Punchblock
 
           it 'sends an error in response to the command' do
             subject.execute_command command
-            command.response.should be == ProtocolError.new('command-not-acceptable', "Did not understand command")
+            command.response.should be == ProtocolError.new.setup('command-not-acceptable', "Did not understand command")
           end
         end
       end