+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # [develop](https://github.com/adhearsion/punchblock)
 
+# [v0.11.0](https://github.com/adhearsion/punchblock/compare/v0.10.0...v0.11.0) - [2012-03-29](https://rubygems.org/gems/punchblock/versions/0.11.0)
+  * Feature: Input & Output components on Asterisk now responds to a Stop command
+  * Feature: started/stopped-speaking events are now handled
+  * Bugfix: Asterisk output component considers an SSML doc w/ a string node w/o spaces to be a filename
+  * Bugfix: `ProtocolError` should behave like a normal exception, just with extra attributes
+
 # [v0.10.0](https://github.com/adhearsion/punchblock/compare/v0.9.2...v0.10.0) - [2012-03-19](https://rubygems.org/gems/punchblock/versions/0.10.0)
   * Feature: app_swift is now supported on Asterisk with a media_engine type of :swift
   * Feature: Asterisk calls now support the Join API